LRQA-64470 Rename
